first commit for sprint 13

+++ controllers/clothingItems.js
@@ -36,28 +36,37 @@ const createItem = (req, res) => {
     });
 };
 
-const deleteItem = (req, res) => {
+const deleteItem = async (req, res) => {
   const { itemId } = req.params;
-
-  ClothingItem.findByIdAndDelete(itemId)
-    .orFail()
-    .then(() => res.status(200).send({ message: "Successfully deleted" }))
-    .catch((err) => {
-      console.error(err);
-      if (err.name === "DocumentNotFoundError") {
-        return res
-          .status(ERROR_CODES.NOT_FOUND)
-          .send({ message: ERROR_MESSAGES.NOT_FOUND });
-      }
-      if (err.name === "CastError") {
-        return res
-          .status(ERROR_CODES.BAD_REQUEST)
-          .send({ message: ERROR_MESSAGES.BAD_REQUEST });
-      }
-      return res
-        .status(ERROR_CODES.SERVER_ERROR)
-        .send({ message: ERROR_MESSAGES.SERVER_ERROR });
+  console.log("clothing item id: ", itemId);
+  try {
+    const item = await ClothingItem.findById(itemId).orFail(() => {
+      const error = new Error("Item ID not found");
+      error.statusCode = 404;
+      throw error;
     });
+
+    if (item.owner.toString() !== req.user._id.toString()) {
+      return res
+        .status(ERROR_CODES.FORBIDDEN)
+        .send({ message: ERROR_MESSAGES.FORBIDDEN });
+    }
+
+    await ClothingItem.findByIdAndDelete(itemId);
+    return res.status(200).send({ message: "Item successfully deleted" });
+  } catch (err) {
+    console.error("deleteItem error name: ", err.name);
+    const statusCode = err.statusCode || 500;
+
+    if (err.name === "CastError") {
+      return res
+        .status(ERROR_CODES.BAD_REQUEST)
+        .send({ message: ERROR_MESSAGES.BAD_REQUEST });
+    }
+    return res
+      .status(statusCode)
+      .send({ message: "An error has occurred on the server." });
+  }
 };
 
 const likeItem = (req, res) => {
